+++ demo/demo1/App.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
-import { createGlobalState } from '../../src/GlobalState/GlobalState';
+import { createGlobalScope, createGlobalState } from '../../src/GlobalState/GlobalState';
 import User from './components/User';
-import UserClass, { UserClassWithGlobalState } from './components/UserClass';
+import UserClass, { UserClassWithGlobalScope } from './components/UserClass';
 import WrappedUserClass from './components/WrappedUserClass'
 import { UserScopeInf } from './types';
+
+const FooGlobalState = createGlobalState('foo', 42);
 
 const userScope: UserScopeInf = {
   name: 'Alex',
@@ -11,37 +13,39 @@ const userScope: UserScopeInf = {
   age: 37,
 }
 
-const UserGlobalState = createGlobalState('user',  userScope, { config: 'config' });
+const UserGlobalScope = createGlobalScope('user',  userScope, { config: 'config' });
 
 function App() {
   return (
-    <UserGlobalState>
-      <div className="App">
-        <b>React Functional Component:</b><br />
-        <User />
+    <FooGlobalState>
+      <UserGlobalScope>
+        <div className="App">
+          <b>React Functional Component:</b><br />
+          <User />
 
-        <hr />
+          <hr />
 
-        <b>User class without Global State:</b><br />
-        <UserClass role='Admin'>
-          Adam & Eva
-        </UserClass>
+          <b>User class without Global State:</b><br />
+          <UserClass role='Admin'>
+            Adam & Eva
+          </UserClass>
 
-        <hr />
+          <hr />
 
-        <b>React Higher-Order Class Component:</b><br />
-        <UserClassWithGlobalState role='Admin'>
-          Adam & Eva
-        </UserClassWithGlobalState>
+          <b>React Higher-Order Class Component:</b><br />
+          <UserClassWithGlobalScope role='Admin'>
+            Adam & Eva
+          </UserClassWithGlobalScope>
 
-        <hr />
+          <hr />
 
-        <b>React Class Component with Wrapper:</b><br />
-        <WrappedUserClass>
-          Adam & Eva
-        </WrappedUserClass>
-      </div>
-    </UserGlobalState>
+          <b>React Class Component with Wrapper:</b><br />
+          <WrappedUserClass>
+            Adam & Eva
+          </WrappedUserClass>
+        </div>
+      </UserGlobalScope>
+    </FooGlobalState>
   );
 }