+++ demo/demo1/components/WrappedUserClass.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
-import { useGlobalState } from '../../../src/GlobalState/GlobalState';
+import { useGlobalScope } from '../../../src/GlobalState/GlobalState';
 import { ConfigScopeInf, UserScopeInf } from '../types';
 import UserClass from './UserClass';
 
 const WrappedUserClass: React.FC<React.PropsWithChildren<{}>> = ({ children = undefined }) => {
-    const configGlobalState = useGlobalState<ConfigScopeInf>('config');
-    const userGlobalState = useGlobalState<UserScopeInf>('user');
+    const configGlobalScope = useGlobalScope<ConfigScopeInf>('config');
+    const userGlobalScope = useGlobalScope<UserScopeInf>('user');
 
     return (
         <UserClass
             role='Admin'
-            userGlobalState={userGlobalState}
-            configGlobalState={configGlobalState}
+            userGlobalScope={userGlobalScope}
+            configGlobalScope={configGlobalScope}
         >
             {children}
         </UserClass>