+++ src/index.ts
@@ -1,9 +1,9 @@
 export {
-    createGlobalState,
-    createMultiGlobalStates,
-    useGlobalState,
-    withGlobalState,
-    GlobalStateType,
+    createGlobalScope,
+    createMultiGlobalScopes,
+    useGlobalScope,
+    withGlobalScope,
+    GlobalScopeType as GlobalStateType,
 } from './GlobalState/GlobalState';
 
 export { Scope, isScope, ScopeInf } from './GlobalState/Scope';