+++ src/variables/isFunction.ts
@@ -1,0 +1,4 @@
+
+export const isFunction = (value: any): value is Function => {
+    return typeof(value) === 'function';
+}